+++ ZigBeeNet/ZDO/Command/SimpleDescriptorResponse.cs
@@ -19,12 +19,12 @@ namespace ZigBeeNet.ZDO.Command
         /**
          * NWKAddrOfInterest command message field.
          */
-        public int NwkAddrOfInterest { get; set; }
+        public ushort NwkAddrOfInterest { get; set; }
 
         /**
          * Length command message fie
          */
-        public int Length { get; set; }
+        public byte Length { get; set; }
 
         /**
          * SimpleDescriptor command message field.
@@ -61,8 +61,8 @@ namespace ZigBeeNet.ZDO.Command
                 return;
             }
 
-            NwkAddrOfInterest = (int)deserializer.Deserialize(ZclDataType.Get(DataType.NWK_ADDRESS));
-            Length = (int)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
+            NwkAddrOfInterest = (ushort)deserializer.Deserialize(ZclDataType.Get(DataType.NWK_ADDRESS));
+            Length = (byte)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
             SimpleDescriptor = (SimpleDescriptor)deserializer.Deserialize(ZclDataType.Get(DataType.SIMPLE_DESCRIPTOR));
         }
